+++ src/library/actions/ProjectActions.ts
@@ -1,7 +1,7 @@
-import { ProjectSettings } from "@commercetools/frontend-domain-types/ProjectSettings";
+import { ProjectSettings } from "shared/types/ProjectSettings";
 import { Event, SDK, ServerOptions } from "@commercetools/frontend-sdk";
 import { GetProjectSettingsAction } from "../../types/actions/ProjectActions";
-import { ComposableCommerceEvents } from "../../types/types";
+import { ComposableCommerceEvents } from "../../types/events/ComposableCommerceEvents";
 
 export type ProjectActions = {
 	getSettings: GetProjectSettingsAction;
@@ -19,7 +19,7 @@ export const getProjectActions = (
 				serverOptions: options.serverOptions,
 			});
 
-			if (!response.isError) {
+			if (response.isError === false) {
 				sdk.trigger(
 					new Event({
 						eventName: "projectSettingsFetched",