+++ src/types/actions/ProjectActions.ts
@@ -1,8 +1,8 @@
-import { ProjectSettings } from "@commercetools/frontend-domain-types/ProjectSettings";
+import { ProjectSettings } from "shared/types/ProjectSettings";
 import { SDKResponse, ServerOptions } from "@commercetools/frontend-sdk";
 
 type GetProjectSettingsAction = (options?: {
 	serverOptions?: ServerOptions;
 }) => Promise<SDKResponse<ProjectSettings>>;
 
-export { GetProjectSettingsAction };
+export { type GetProjectSettingsAction };